requirements.txt
plaintext
allure-pytest==2.13.5
allure-python-commons==2.13.5
atomicwrites==1.4.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
Faker==26.0.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
iniconfig==2.0.0
lxml==5.2.2
namedlist==1.8
outcome==1.3.0.post0
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
py==1.11.0
pycparser==2.22
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.2
selenium==4.21.0
setuptools==70.3.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
toml==0.10.2
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
urllib3==2.2.1
webdriver-manager==4.0.1
wheel==0.43.0
wsproto==1.2.0

